% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlebuhanec.0325-package.R
\docType{package}
\name{hlebuhanec.0325-package}
\alias{hlebuhanec.0325}
\alias{hlebuhanec.0325-package}
\title{hlebuhanec.0325:}
\description{

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
