% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{export}
\usage{
plot_cor(
  x,
  y = NULL,
  method = "spearman",
  color = "red",
  alpha = 0.5,
  cex = 1,
  pch_size = cex * 3,
  xlab = NULL,
  ylab = NULL,
  legend = NULL,
  dec = 1,
  wrap = 20,
  model = "lm"
)
}
\description{
export
}
